requirements.txt
# Web框架
Flask==3.0.0
Flask-CORS==4.0.0

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
pdf2image==1.16.3
pytesseract==0.3.10

# AI服务
requests==2.31.0

# 数据处理
pandas==2.1.4
openpyxl==3.1.2

# 网络爬虫
beautifulsoup4==4.12.2
selenium==4.15.2

# 工具库
python-dotenv==1.0.0